+++ src/lib/st-toggle-buttons/st-toggle-buttons.component.spec.ts
@@ -14,7 +14,7 @@ import { By } from '@angular/platform-browser';
 import { RouterTestingModule } from '@angular/router/testing';
 
 import { StToggleButtonsComponent } from './st-toggle-buttons.component';
-import { StToggleButton } from './st-toggle-buttons.interface';
+import { StToggleButton, StParsedToggleButton } from './st-toggle-buttons.interface';
 
 
 let comp: StToggleButtonsComponent;
@@ -22,12 +22,24 @@ let fixture: ComponentFixture<StToggleButtonsComponent>;
 let de: DebugElement;
 
 let tabs: StToggleButton[] = [
-   { label: 'toogle1', active: true, number: 0 },
-   { label: 'toogle2', active: false, number: 1 },
-   { label: 'toogle3', active: false, number: 2 }
+   { label: 'tÇ*´oogle 1', active: true, number: 0 },
+   { label: 'toogle 2', active: false, number: 1 },
+   { label: 'too,gle 3.,-', active: false, number: 2 }
 ];
+let parsedTabs: StParsedToggleButton[] = [
+   { label: 'tÇ*´oogle 1', active: true, number: 0, trimLabel: 'toogle1' },
+   { label: 'toogle 2', active: true, number: 1, trimLabel: 'toogle2' },
+   { label: 'too,gle 3.,-', active: false, number: 2, trimLabel: 'toogle3' }
+];
+const enum trimLabels {
+   label1 = 'toogle1',
+   label2 = 'toogle2',
+   label3 = 'toogle3'
+}
 let description: string = 'toogle test';
 let qaTag: string = 'toogle-test';
+const tabActiveCSSClass = 'st-toggle-buttons__button--active';
+const realisticClick = new Event('click', { bubbles: true });
 
 describe('StToggleButtonsComponent', () => {
 
@@ -51,7 +63,7 @@ describe('StToggleButtonsComponent', () => {
    it('should init correctly', () => {
       fixture.detectChanges();
       let desc: DebugElement = fixture.debugElement.query(By.css('.st-toggle-buttons__description'));
-      let options: DebugElement[] = fixture.debugElement.queryAll(By.css('.st-toggle-buttons__tab'));
+      let options: DebugElement[] = fixture.debugElement.queryAll(By.css('.st-toggle-buttons__button'));
 
       expect(desc).toBeDefined();
       expect(desc.nativeElement).toBeDefined();
@@ -65,19 +77,19 @@ describe('StToggleButtonsComponent', () => {
    it('should change active tab on click', () => {
       fixture.detectChanges();
 
-      let option1: DebugElement = fixture.debugElement.query(By.css(`#${qaTag}-${tabs[0].label}`));
-      let option3: DebugElement = fixture.debugElement.query(By.css(`#${qaTag}-${tabs[2].label}`));
+      let option1: DebugElement = fixture.debugElement.query(By.css(`#${qaTag}-${trimLabels.label1}`));
+      let option3: DebugElement = fixture.debugElement.query(By.css(`#${qaTag}-${trimLabels.label3}`));
 
       expect(option1).toBeDefined();
       expect(option3).toBeDefined();
-      expect(option1.classes['st-toggle-buttons__tab--active']).toBeTruthy();
-      expect(option3.classes['st-toggle-buttons__tab--active']).toBeFalsy();
+      expect(option1.classes[tabActiveCSSClass]).toBeTruthy();
+      expect(option3.classes[tabActiveCSSClass]).toBeFalsy();
 
-      option3.nativeElement.dispatchEvent(new Event('click'));
+      option3.nativeElement.dispatchEvent(realisticClick);
       fixture.detectChanges();
 
-      expect(option1.classes['st-toggle-buttons__tab--active']).toBeFalsy();
-      expect(option3.classes['st-toggle-buttons__tab--active']).toBeTruthy();
+      expect(option1.classes[tabActiveCSSClass]).toBeFalsy();
+      expect(option3.classes[tabActiveCSSClass]).toBeTruthy();
    });
 
    it('Should emit an event when active tab is clicked', () => {
@@ -86,17 +98,17 @@ describe('StToggleButtonsComponent', () => {
 
       fixture.detectChanges();
 
-      let option2: DebugElement = fixture.debugElement.query(By.css(`#${qaTag}-${tabs[1].label}`));
+      let option2: DebugElement = fixture.debugElement.query(By.css(`#${qaTag}-${trimLabels.label2}`));
 
       expect(option2).toBeDefined();
-      expect(option2.classes['st-toggle-buttons__tab--active']).toBeFalsy();
+      expect(option2.classes[tabActiveCSSClass]).toBeFalsy();
 
-      option2.nativeElement.dispatchEvent(new Event('click'));
+      option2.nativeElement.dispatchEvent(realisticClick);
       fixture.detectChanges();
 
-      expect(option2.classes['st-toggle-buttons__tab--active']).toBeTruthy();
+      expect(option2.classes[tabActiveCSSClass]).toBeTruthy();
       expect(responseFunction).toHaveBeenCalled();
       expect(responseFunction).toHaveBeenCalledTimes(1);
-      expect(responseFunction).toHaveBeenCalledWith(tabs[1]);
+      expect(responseFunction).toHaveBeenCalledWith(parsedTabs[1]);
    });
 });